split into different states (unstable)
the main idea is to split the fumen editor into the display, book, and history.
the book (mostly) only be updated through the display, the display renders either from the book or after the displayState is modified, and historyState stores any modified bookState.
this can hide a lot of automatic interactions between the states that were previously explicitly carried out
it's also easier to separate what scale the state is changed